refactor(firmware): dedupe the Android/JVM download byte-copy loop
AndroidFirmwareFileHandler and JvmFirmwareFileHandler each hand-rolled
an identical byte-copy-with-progress loop (including the same
cancellation check and incomplete-download detection), a drift risk on
a critical OTA path -- the same rationale already documented for why
extractZipEntriesBounded lives in the shared jvmAndroidMain source set.

Extracted downloadResponseToFile() into that same source set. Kept the
existing incremental-streaming approach rather than switching to
ktor's onDownload + .body(): ByteArray, since that would buffer an
entire firmware image (tens of MB) in memory to drop a manual loop --
not worth it for files this size.
