+++ feature/firmware/src/androidMain/kotlin/org/meshtastic/feature/firmware/AndroidFirmwareFileHandler.kt
@@ -23,13 +23,8 @@ import com.eygraber.uri.toAndroidUri
 import io.ktor.client.HttpClient
 import io.ktor.client.request.get
 import io.ktor.client.request.head
-import io.ktor.client.statement.bodyAsChannel
 import io.ktor.client.statement.bodyAsText
-import io.ktor.http.contentLength
 import io.ktor.http.isSuccess
-import io.ktor.utils.io.jvm.javaio.toInputStream
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.isActive
 import kotlinx.coroutines.withContext
 import org.koin.core.annotation.Single
 import org.meshtastic.core.common.util.CommonUri
@@ -42,8 +37,6 @@ import java.io.InputStream
 import java.net.URI
 import java.util.zip.ZipEntry
 import java.util.zip.ZipInputStream
-
-private const val DOWNLOAD_BUFFER_SIZE = 8192
 
 /**
  * Helper class to handle file operations related to firmware updates, such as downloading, copying from URI, and
@@ -98,34 +91,9 @@ class AndroidFirmwareFileHandler(private val context: Context, private val clien
                 return@withContext null
             }
 
-            val body = response.bodyAsChannel()
-            val contentLength = response.contentLength() ?: -1L
-
             if (!tempDir.exists()) tempDir.mkdirs()
-
             val targetFile = java.io.File(tempDir, fileName)
-
-            body.toInputStream().use { input ->
-                java.io.FileOutputStream(targetFile).use { output ->
-                    val buffer = ByteArray(DOWNLOAD_BUFFER_SIZE)
-                    var bytesRead: Int
-                    var totalBytesRead = 0L
-
-                    while (input.read(buffer).also { bytesRead = it } != -1) {
-                        if (!isActive) throw CancellationException("Download cancelled")
-
-                        output.write(buffer, 0, bytesRead)
-                        totalBytesRead += bytesRead
-
-                        if (contentLength > 0) {
-                            onProgress(totalBytesRead.toFloat() / contentLength)
-                        }
-                    }
-                    if (contentLength != -1L && totalBytesRead != contentLength) {
-                        throw IOException("Incomplete download: expected $contentLength bytes, got $totalBytesRead")
-                    }
-                }
-            }
+            downloadResponseToFile(response, targetFile, onProgress)
             targetFile.toFirmwareArtifact()
         }
